requirements.txt
discord.py
feedparser
requests
googletrans==4.0.0-rc1
python-dotenv
